requirements.txt
alembic==1.4.3
APScheduler==3.6.3
backcall==0.2.0
click==7.1.2
colorama==0.4.4
colorhash==1.0.2
configparser==5.0.1
decorator==4.4.2
fire==0.3.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.1
jupyter-client==6.1.7
jupyter-core==4.6.3
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.4
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
psutil==5.7.3
Pygments==2.7.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
pywin32==228
PyYAML==5.3.1
pyzmq==19.0.2
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.20
termcolor==1.1.0
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
